requirements.txt
fastmcp>=2.13.1
python-dotenv
requests
pydantic
PyJWT
httpx
